requirements.txt
# TikTok uploader
tiktokautouploader

# Pinterest uploader (browser automation)
zendriver
python-dotenv

# YouTube uploader (browser automation)
playwright
